Split lcov parser into a focused module
